requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
discord.py==2.6.3
dotenv==0.9.9
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
pycparser==2.23
PyNaCl==1.5.0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
spotipy==2.25.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
